requirements.txt
langchain
langgraph
langchain-core
langchain-community
python-dotenv
langchain-openai
langchain-groq
arxiv
wikipedia
langgraph-cli[inmem]
chromadb
ipykernel